sprint1/step1 правка файловой структуры проекта, перенос key в списках
