+++ src/components/app/app.module.css
@@ -1,4 +1,4 @@
-.main-container {
+.main_container {
   display: grid;
   grid-template-columns: repeat(2, 1fr);
   column-gap: 40px;
